:sparkles: Implementing and :goal_net: Error Handling :sparkles: Implementing the DataBaseConnectionException, DragAlreadyExistsException, NothingToDeleteException :goal_net: Handling the DataBaseConnectionException, DragAlreadyExistsException, NothingToDeleteException :sparkles: Implementing Message modeling class to return REST messages

+++ src/main/java/br/inatel/dragrace/service/DragService.java
@@ -5,15 +5,16 @@ import br.inatel.dragrace.adapter.dto.CarDto;
 import br.inatel.dragrace.adapter.dto.CarRequestDto;
 import br.inatel.dragrace.controller.dto.DragDto;
 import br.inatel.dragrace.controller.form.DragForm;
-import br.inatel.dragrace.exception.CarDataApiConnectionException;
-import br.inatel.dragrace.exception.CarNotFoundException;
+import br.inatel.dragrace.exception.*;
 import br.inatel.dragrace.mapper.DragMapper;
 import br.inatel.dragrace.model.Drag;
 import br.inatel.dragrace.model.SpeedWinner;
 import br.inatel.dragrace.model.TimeWinner;
+import br.inatel.dragrace.model.rest.Message;
 import br.inatel.dragrace.repository.DragRepository;
 import br.inatel.dragrace.repository.SpeedWinnerRepository;
 import br.inatel.dragrace.repository.TimeWinnerRepository;
+import org.hibernate.exception.JDBCConnectionException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClientException;
@@ -59,27 +60,55 @@ public class DragService {
 
     }
     public DragDto newDrag(CarRequestDto carRequestDto, DragForm dragForm) {
-        CarDto carDto = getCarFromCarData(carRequestDto);
-        Drag drag = new Drag(carDto, dragForm);
-        dragRepository.save(drag);
-        return DragMapper.toDragDto(drag);
+        try {
+            Drag verifierDrag = dragRepository.findByDriver(dragForm.getDriver());
+            if (verifierDrag != null) {
+                throw new DragAlreadyExistsException(dragForm.getDriver());
+            } else {
+                CarDto carDto = getCarFromCarData(carRequestDto);
+                Drag drag = new Drag(carDto, dragForm);
+                dragRepository.save(drag);
+                return DragMapper.toDragDto(drag);
+            }
+        } catch (JDBCConnectionException jdbcConnectionException){
+            throw new DataBaseConnectionException(jdbcConnectionException);
+        }
     }
     public List<DragDto> listALlDrags() {
-        List<Drag> drags = dragRepository.findAll();
-        List<DragDto> dragDtos = DragMapper.toListDragDto(drags);
-        return dragDtos;
+        try {
+            List<Drag> drags = dragRepository.findAll();
+            List<DragDto> dragDtos = DragMapper.toListDragDto(drags);
+            return dragDtos;
+        } catch (JDBCConnectionException jdbcConnectionException){
+            throw new DataBaseConnectionException(jdbcConnectionException);
+        }
     }
 
     public void setWinners() {
-        Drag raceTimeWinner = dragRepository.receTimeWinner();
-        Drag raceSpeedWinner = dragRepository.receSpeedWinner();
-        TimeWinner timeWinner = new TimeWinner(raceTimeWinner);
-        SpeedWinner speedWinner = new SpeedWinner(raceSpeedWinner);
-        timeWinnerRepository.save(timeWinner);
-        speedWinnerRepository.save(speedWinner);
+        try {
+            Drag raceTimeWinner = dragRepository.receTimeWinner();
+            Drag raceSpeedWinner = dragRepository.receSpeedWinner();
+            TimeWinner timeWinner = new TimeWinner(raceTimeWinner);
+            SpeedWinner speedWinner = new SpeedWinner(raceSpeedWinner);
+            timeWinnerRepository.save(timeWinner);
+            speedWinnerRepository.save(speedWinner);
+        } catch (JDBCConnectionException jdbcConnectionException){
+            throw new DataBaseConnectionException(jdbcConnectionException);
+        }
     }
 
-    public void resetRace() {
-        dragRepository.deleteAll();
+    public Message resetRace() {
+        try {
+            List<Drag> drags = dragRepository.findAll();
+            if(drags.isEmpty()){
+                throw new NothingToDeleteException();
+            }else {
+                dragRepository.deleteAll();
+                return new Message("The race was reseted successfully");
+            }
+        } catch (JDBCConnectionException jdbcConnectionException){
+            throw new DataBaseConnectionException(jdbcConnectionException);
+        }
+
     }
 }
